Operators Assignment
Perform these actions and create a console app that includes the following:

Create an Employee class with Id, FirstName and LastName properties.
In the Employee class, overload the “==” operator so it checks if two Employee objects are equal by comparing their Id property. Remember that comparison operators must be overloaded in pairs.

Add comments to each line or block of your code to explain what it does exactly, so that another developer could read and understand your code.