Input Assignment
Input Assignment
Perform these actions and create a console app that includes the following:

Asks the user for a number.

Logs that number to a text file.

Prints the text file back to the user.

Add comments to each line or block of your code to explain what it does exactly, so that another developer could read and understand your code.

Upload your code to GitHub.