Datetime Assignment
Datetime Assignment

Perform these actions and create a console app that includes the following:

Prints the current date and time to the console.

Asks the user for a number.

Prints to the console the exact time it will be in X hours, X being the number the user entered in step 2.

Add comments to each line or block of your code to explain what it does exactly, so that another developer could read and understand your code.

Upload your code to GitHub.